Refactoring tcpClient and add new module
